Add an undocumented PL_STANDARD_CLOCK flag to control the default clock implementation.
When set to 1, it forces the usage of the slower but always correct high resolution clock
(default is 0) as some old processors have a non monotonic behavior of the fast `rdtsc`
instruction.
It is used by default on Python to make it more reliable, as the effect on speed is not visible there.